migration for the rest of entity, voting, story, etc

+++ socket/src/_db/migrations/1680838020893-CreateUsersTable.ts
@@ -8,7 +8,7 @@ export class CreateUsersTable1680838020893 implements MigrationInterface {
         columns: [
           {
             name: 'id',
-            type: 'integer',
+            type: 'bigint',
             isPrimary: true,
             isGenerated: true,
             generationStrategy: 'increment',
